Fix missing argument to mkTeacher function
When handling the "set" method in the
`closures/teacher.nix` example in the
definition for `mkTeacher`, when we receive
a `key` value whose behavior we don't override,
there was a missing final argument for the `field`
parameter of the `mkTeacher` function.

We also add some code that exercises this code branch.